+++ primitive/ray.go
@@ -4,7 +4,7 @@ type Ray struct {
 	Origin, Direction Vector
 }
 
-func (r Ray) Point(t float64) Vector {
+func (r Ray) Point(t float32) Vector {
 	b := r.Direction.MulScalar(t)
 	return r.Origin.Add(b)
 }
