+++ scene/hitable.go
@@ -3,7 +3,7 @@ package scene
 import "github.com/ruegerj/raytracing/primitive"
 
 type Hit struct {
-	Distance float64
+	Distance float32
 	Point    primitive.Vector
 	Normal   primitive.Vector
 	Color    primitive.ScalarColor
